Fix menu preview rendering
The menu preview was off in a few ways. The parent relative position
for the menu items was incorrect, causing them to grab the wrong part
of their parent's texture.

The menu title's parent was unset, which means it crashed when
parentrealtive was used.

The borders were drawn in the wrong places since the menu items were not
drawn in the right places, and the menu background was being drawn into
the preview window when it should only appear as a result of the menu items
reusing the texture in their own. Lastly, the menu background's height
needed to be fixed to include 1 border (between the title and the items).

This fixes the crash reported in Bug 4659.