Stop using libglade
GtkOptionMenu -> GtkComboBoxText move is to make porting
to gtk3 easier (this needs gtk+ 2.24), most of the pixbuf
changes in src/preview.c for the same reason. Other changes
should not go beyond 2.18.

The changes in packing in glade file are also for the sake
of gtk3 - they shouldn't affect gtk2 (hbox51 removal was
mostly to make things cleaner - there didn't seem to be a
reason for extra container there).

The added catch is that http://savannah.gnu.org/bugs/?34506
was only fixed in gettext 0.18.3 (just a few weeks old) and
there's a little problem with gtk2 version of glade
(http://bugzilla.gnome.org/show_bug.cgi?id=689667). The ui
file itself nevertheless works.